[Feat] STAMPIT-139 마이페이지 스와이프하여 탭 이동
