+++ StampIt-Project/StampIt-Project/Presentation/MyPage/ViewController/MyPageViewController.swift
@@ -14,19 +14,33 @@ final class MyPageViewController: UIViewController {
     
     // MARK: - Properties
     
-    private var viewModel: MyPageViewModel
-    private var container: DIContainer
+    private let viewModel: MyPageViewModel
+    private let container: DIContainer
     private let disposeBag = DisposeBag()
     
     // MARK: - UI Components
-
+    
+    private lazy var viewControllers: [UIViewController] = [
+        container.makeStampBoardViewController(),
+        container.makeProfileViewController()
+    ]
+    
+    private lazy var pageViewController = UIPageViewController(
+        transitionStyle: .scroll,
+        navigationOrientation: .horizontal
+    ).then {
+        $0.setViewControllers(
+            [viewControllers[0]],
+            direction: .forward,
+            animated: false
+        )
+    }
+    
     private let navigationBar = DefaultNavigationBar(
         .segmentedControlTabs(
             tab1: TabType.stampBoard.title,
             tab2: TabType.profile.title
         ))
-    private let stampBoardView = StampBoardTab()
-    private let profileView = ProfileTab()
     
     // MARK: - Initializer, Deinit, requiered
     
@@ -47,30 +61,21 @@ final class MyPageViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        viewModel.action.accept(.viewDidLoad)
         setStyle()
         setHierarchy()
         setLayout()
         setDelegate()
         setDataSource()
-        setAction()
         bind()
-    }
-    
-    // 화면이 나타날 때마다 데이터 새로고침
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        viewModel.action.accept(.viewDidLoad)
     }
     
     // MARK: - Bind
     
     private func bind() {
-        
         navigationBar.tabTapped
             .bind(with: self) { owner, type in
                 owner.viewModel.action.accept(
-                    .tabButtonTapped(TabType(rawValue: type.rawValue)!)
+                    .tabChanged(TabType(rawValue: type.rawValue)!)
                 )
             }.disposed(by: disposeBag)
     
@@ -79,68 +84,29 @@ final class MyPageViewController: UIViewController {
                 owner.navigationBar.updateTabTitleColor(selected: tab)
                 owner.updateSelectedTab(selected: tab)
             }.disposed(by: disposeBag)
-        
-        Observable.combineLatest(
-            viewModel.state.stickerSummary,
-            viewModel.state.stickers
-        )
-        .bind(with: self) { owner, combined in
-            let (summary, stickers) = combined
-            owner.updateSnapshot(summary: summary, stickers: stickers)
-        }.disposed(by: disposeBag)
-        
-        viewModel.state.user
-            .compactMap { $0 }
-            .bind(with: self) { owner, user in
-                owner.profileView.setUser(user)
-            }.disposed(by: disposeBag)
-        
-        viewModel.state.alertMessage
-            .bind(with: self) { owner, message in
-                owner.showAlert(title: "1인 그룹에서는 그룹 탈퇴가 지원되지 않아요.", message: message)
-            }.disposed(by: disposeBag)
-        
-        // 로그인 화면으로 이동
-        viewModel.state.shouldNavigateToLogin
-            .bind(with: self) { owner, _ in
-                // 토스트가 이미 떠있으면 1.5초 뒤에 전환
-                DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
-                    owner.navigateToLogin()
-                }
-            }.disposed(by: disposeBag)
-
-        // 확인 다이얼로그
-        viewModel.state.shouldShowConfirmAlert
-            .bind(with: self) { owner, alertData in
-                let (title, message, action) = alertData
-                owner.showConfirmAlert(title: title, message: message, confirmAction: action)
-            }.disposed(by: disposeBag)
-        
-        // 토스트 메시지
-        viewModel.state.toastMessage
-            .bind(with: self) { owner, message in
-                owner.showToast(message: message, type: .success)
-            }.disposed(by: disposeBag)
     }
     
     // MARK: - Style Helper
     
     private func setStyle() {
         view.backgroundColor = .white
-        stampBoardView.isHidden = false
-        profileView.isHidden = true
         navigationController?.setNavigationBarHidden(true, animated: false)
     }
     
     // MARK: - Hierarchy Helper
     
     private func setHierarchy() {
+        /// 자식 ViewController 로 설정
+        addChild(pageViewController)
+        
         [
             navigationBar,
-            stampBoardView,
-            profileView
+            pageViewController.view
         ]
             .forEach { view.addSubview($0) }
+        
+        /// 자식 ViewController 에 didMove 호출
+        pageViewController.didMove(toParent: self)
     }
 
     // MARK: - Layout Helper
@@ -150,146 +116,80 @@ final class MyPageViewController: UIViewController {
             $0.top.directionalHorizontalEdges.equalTo(view.safeAreaLayoutGuide)
         }
         
-        stampBoardView.snp.makeConstraints {
+        pageViewController.view.snp.makeConstraints {
             $0.top.equalTo(navigationBar.snp.bottom)
-            $0.directionalHorizontalEdges.bottom.equalToSuperview()
-        }
-        
-        profileView.snp.makeConstraints {
-            $0.top.equalTo(navigationBar.snp.bottom)
-            $0.directionalHorizontalEdges.bottom.equalToSuperview()
+            $0.bottom.directionalHorizontalEdges.equalToSuperview()
         }
     }
     
     // MARK: - Delegate Helper
     
     private func setDelegate() {
-        profileView.tableView.delegate = self
+        pageViewController.delegate = self
     }
 
     // MARK: - DataSource Helper
     
     private func setDataSource() {
-        profileView.tableView.dataSource = self
-    }
-
-    // MARK: - Snapshot
-    
-    private func updateSnapshot(
-        summary: (collected: Int, completed: Int),
-        stickers: [Sticker]
-    ) {
-        var snapshot = NSDiffableDataSourceSnapshot<StampBoardSection, StampBoardItem>()
-        snapshot.appendSections([.summary])
-        snapshot.appendItems(
-            [.summary(
-                collected: summary.collected,
-                completed: summary.completed
-            )],
-            toSection: .summary
-        )
-        
-        snapshot.appendSections([.defaultBoard])
-        snapshot.appendItems(
-            stickers.map { .stickers($0) } ,
-            toSection: .defaultBoard
-        )
-        stampBoardView.stickerBoardDataSource.apply(snapshot, animatingDifferences: false)
-    }
-    
-    // MARK: - Button Action Helper
-    
-    private func setAction() {
-        profileView.setButtonAction(target: self, action: #selector(tappedEditButton))
-    }
-    
-    @objc private func tappedEditButton() {
-        let user = viewModel.state.user.value
-        guard let user else { return }
-        let viewController = container.makeEditProfileViewController(user: user)
-        navigationController?.pushViewController(viewController, animated: true)
+        pageViewController.dataSource = self
     }
     
     // MARK: - Methods
     
+    /// 탭 클릭시 ViewController 전환
     private func updateSelectedTab(selected: TabType) {
-        switch selected {
-        case .stampBoard:
-            stampBoardView.isHidden = false
-            profileView.isHidden = true
-            
-        case .profile:
-            stampBoardView.isHidden = true
-            profileView.isHidden = false
-        }
-    }
-    
-    /// 일반 알림 다이얼로그
-    private func showAlert(title: String, message: String) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "확인", style: .default))
-        present(alert, animated: true)
-    }
-    
-    /// 확인/취소 다이얼로그
-    private func showConfirmAlert(title: String, message: String, confirmAction: @escaping () -> Void) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        guard let currenctVC = pageViewController.viewControllers?.first,
+              let currentIndex = viewControllers.firstIndex(of: currenctVC) else { return }
         
-        alert.addAction(UIAlertAction(title: "취소", style: .cancel))
-        alert.addAction(UIAlertAction(title: "확인", style: .destructive) { _ in
-            confirmAction()
-        })
+        let newIndex = selected.index
         
-        present(alert, animated: true)
-    }
-    
-    /// TODO: 리더 전용 메뉴 표시 (향후 기능 완성 후 활성화)
-     /*
-    private func showLeaderOptionsAlert() {
-        let alert = UIAlertController(title: "그룹장 옵션", message: "그룹을 떠나려면 먼저 다음 작업을 수행해주세요:", preferredStyle: .actionSheet)
-        alert.addAction(UIAlertAction(title: "그룹장 위임하기", style: .default) { _ in // TODO: 리더 위임 화면으로 이동
-        })
-        alert.addAction(UIAlertAction(title: "멤버 관리하기", style: .default) { _ in // TODO: 멤버 관리 화면으로 이동
-        })
-        alert.addAction(UIAlertAction(title: "계정 탈퇴 (그룹 삭제)", style: .destructive) { _ in self.deleteAccount() })
-        alert.addAction(UIAlertAction(title: "취소", style: .cancel))
-        present(alert, animated: true)
-    }
-    */
-    
-    /// 로그인 화면으로 이동
-    private func navigateToLogin() {
-        UserCache.shared.clearCache()
+        /// 같은 탭이면 무시
+        guard newIndex != currentIndex else { return }
         
-         // UserDefaults에서 로그인 관련 정보 삭제 (필요시)
-         UserDefaults.standard.removeObject(forKey: "userToken")
-         UserDefaults.standard.removeObject(forKey: "lastLoginDate")
+        let direction: UIPageViewController.NavigationDirection = newIndex > currentIndex ? .forward : .reverse
         
-        let loginViewModel = DIContainer.shared.makeLoginViewModel()
-        let loginVC = LoginViewController(viewModel: loginViewModel, container: container)
-        let navController = UINavigationController(rootViewController: loginVC)
-        
-        WindowTransitionManager.shared.changeRootViewController(to: navController)
-    }
-    
-    // 외부에서 쓸 수 있는 메서드로 액션 전달 (public/internal)
-    func leaveGroup() {
-        viewModel.action.accept(.leaveGroupButtonTapped)
-    }
-
-    func deleteAccount() {
-        viewModel.action.accept(.deleteAccountButtonTapped)
-    }
-    
-    func logOut() {
-        viewModel.action.accept(.logoutButtonTapped)
+        pageViewController.setViewControllers(
+            [viewControllers[newIndex]],
+            direction: direction,
+            animated: true
+        )
     }
 }
 
-// 토스트 메시지(2초)
-extension MyPageViewController {
-    func showToast(message: String, type: ToastType = .success, duration: TimeInterval = 2.0) {
-        let toastView = ToastView()
-        toastView.show(in: self.view, duration: duration, message: message, type: type)
+/// Swiping 하여 ViewController 전환
+extension MyPageViewController: UIPageViewControllerDelegate {
+    func pageViewController(
+        _ pageViewController: UIPageViewController,
+        didFinishAnimating finished: Bool,
+        previousViewControllers: [UIViewController],
+        transitionCompleted completed: Bool
+    ) {
+        if completed,
+           let newVC = pageViewController.viewControllers?.first,
+           let newIndex = viewControllers.firstIndex(of: newVC) {
+            viewModel.action.accept(
+                .tabChanged(TabType.index(newIndex))
+            )
+        }
+    }
+}
+
+extension MyPageViewController: UIPageViewControllerDataSource {
+    func pageViewController(
+        _ pageViewController: UIPageViewController,
+        viewControllerBefore viewController: UIViewController
+    ) -> UIViewController? {
+        guard let currentIndex = viewControllers.firstIndex(of: viewController),
+              currentIndex > 0 else { return nil }
+        return viewControllers[currentIndex - 1]
+    }
+    
+    func pageViewController(
+        _ pageViewController: UIPageViewController,
+        viewControllerAfter viewController: UIViewController
+    ) -> UIViewController? {
+        guard let currentIndex = viewControllers.firstIndex(of: viewController),
+              currentIndex < viewControllers.count - 1 else { return nil }
+        return viewControllers[currentIndex + 1]
     }
 }
