+++ StampIt-Project/StampIt-Project/Presentation/MyPage/ViewController/ProfileViewController.swift
@@ -1,0 +1,215 @@
+//
+//  ProfileTabViewController.swift
+//  StampIt-Project
+//
+//  Created by kingj on 6/24/25.
+//
+
+import UIKit
+import Then
+import SnapKit
+import RxSwift
+
+final class ProfileViewController: UIViewController {
+    
+    // MARK: - Properties
+    
+    private var viewModel: ProfileViewModel
+    private var container: DIContainer
+    private let disposeBag = DisposeBag()
+    
+    // MARK: - UI Components
+
+    private let profileView = ProfileTab()
+    
+    // MARK: - Initializer, Deinit, requiered
+    
+    init(
+        viewModel: ProfileViewModel,
+        container: DIContainer
+    ) {
+        self.viewModel = viewModel
+        self.container = container
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - View Life Cycle
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setStyle()
+        setHierarchy()
+        setLayout()
+        setDelegate()
+        setDataSource()
+        setAction()
+        bind()
+    }
+    
+    // 화면이 나타날 때마다 데이터 새로고침
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        // user 다시 불러오기?
+    }
+    
+    // MARK: - Bind
+    
+    private func bind() {
+        viewModel.state.user
+            .compactMap { $0 }
+            .bind(with: self) { owner, user in
+                owner.profileView.setUser(user)
+            }.disposed(by: disposeBag)
+        
+        viewModel.state.alertMessage
+            .bind(with: self) { owner, message in
+                owner.showAlert(title: "1인 그룹에서는 그룹 탈퇴가 지원되지 않아요.", message: message)
+            }.disposed(by: disposeBag)
+        
+        // 로그인 화면으로 이동
+        viewModel.state.shouldNavigateToLogin
+            .bind(with: self) { owner, _ in
+                // 토스트가 이미 떠있으면 1.5초 뒤에 전환
+                DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
+                    owner.navigateToLogin()
+                }
+            }.disposed(by: disposeBag)
+
+        // 확인 다이얼로그
+        viewModel.state.shouldShowConfirmAlert
+            .bind(with: self) { owner, alertData in
+                let (title, message, action) = alertData
+                owner.showConfirmAlert(title: title, message: message, confirmAction: action)
+            }.disposed(by: disposeBag)
+        
+        // 토스트 메시지
+        viewModel.state.toastMessage
+            .bind(with: self) { owner, message in
+                owner.showToast(message: message, type: .success)
+            }.disposed(by: disposeBag)
+    }
+    
+    // MARK: - Style Helper
+    
+    private func setStyle() {
+        view.backgroundColor = .white
+        navigationController?.setNavigationBarHidden(true, animated: false)
+    }
+    
+    // MARK: - Hierarchy Helper
+    
+    private func setHierarchy() {
+        [
+            profileView
+        ]
+            .forEach { view.addSubview($0) }
+    }
+
+    // MARK: - Layout Helper
+    
+    private func setLayout() {
+        profileView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
+    
+    // MARK: - Delegate Helper
+    
+    private func setDelegate() {
+        profileView.tableView.delegate = self
+    }
+
+    // MARK: - DataSource Helper
+    
+    private func setDataSource() {
+        profileView.tableView.dataSource = self
+    }
+    
+    // MARK: - Button Action Helper
+    
+    private func setAction() {
+        profileView.setButtonAction(target: self, action: #selector(tappedEditButton))
+    }
+    
+    @objc private func tappedEditButton() {
+        let user = viewModel.state.user.value
+        guard let user else { return }
+        let viewController = container.makeEditProfileViewController(user: user)
+        navigationController?.pushViewController(viewController, animated: true)
+    }
+    
+    // MARK: - Methods
+    
+    /// 일반 알림 다이얼로그
+    private func showAlert(title: String, message: String) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        alert.addAction(UIAlertAction(title: "확인", style: .default))
+        present(alert, animated: true)
+    }
+    
+    /// 확인/취소 다이얼로그
+    private func showConfirmAlert(title: String, message: String, confirmAction: @escaping () -> Void) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        
+        alert.addAction(UIAlertAction(title: "취소", style: .cancel))
+        alert.addAction(UIAlertAction(title: "확인", style: .destructive) { _ in
+            confirmAction()
+        })
+        
+        present(alert, animated: true)
+    }
+    
+    /// TODO: 리더 전용 메뉴 표시 (향후 기능 완성 후 활성화)
+     /*
+    private func showLeaderOptionsAlert() {
+        let alert = UIAlertController(title: "그룹장 옵션", message: "그룹을 떠나려면 먼저 다음 작업을 수행해주세요:", preferredStyle: .actionSheet)
+        alert.addAction(UIAlertAction(title: "그룹장 위임하기", style: .default) { _ in // TODO: 리더 위임 화면으로 이동
+        })
+        alert.addAction(UIAlertAction(title: "멤버 관리하기", style: .default) { _ in // TODO: 멤버 관리 화면으로 이동
+        })
+        alert.addAction(UIAlertAction(title: "계정 탈퇴 (그룹 삭제)", style: .destructive) { _ in self.deleteAccount() })
+        alert.addAction(UIAlertAction(title: "취소", style: .cancel))
+        present(alert, animated: true)
+    }
+    */
+    
+    /// 로그인 화면으로 이동
+    private func navigateToLogin() {
+        UserCache.shared.clearCache()
+        
+         // UserDefaults에서 로그인 관련 정보 삭제 (필요시)
+         UserDefaults.standard.removeObject(forKey: "userToken")
+         UserDefaults.standard.removeObject(forKey: "lastLoginDate")
+        
+        let loginViewModel = DIContainer.shared.makeLoginViewModel()
+        let loginVC = LoginViewController(viewModel: loginViewModel, container: container)
+        let navController = UINavigationController(rootViewController: loginVC)
+        
+        WindowTransitionManager.shared.changeRootViewController(to: navController)
+    }
+    
+    // 외부에서 쓸 수 있는 메서드로 액션 전달 (public/internal)
+    func leaveGroup() {
+        viewModel.action.accept(.leaveGroupButtonTapped)
+    }
+
+    func deleteAccount() {
+        viewModel.action.accept(.deleteAccountButtonTapped)
+    }
+    
+    func logOut() {
+        viewModel.action.accept(.logoutButtonTapped)
+    }
+}
+
+// 토스트 메시지(2초)
+extension ProfileViewController {
+    func showToast(message: String, type: ToastType = .success, duration: TimeInterval = 2.0) {
+        let toastView = ToastView()
+        toastView.show(in: self.view, duration: duration, message: message, type: type)
+    }
+}
